Adds skip functions to sync
Offers a third option between force sync everything and sync nothing if
the first opds item return matches an item in our library.

With OPDS server entries where you have saved the result of a search as
your sync (against tags, for example), and the latest tagged item isn't
necessarily the latest added item, this lets you still sync successfully
without being forced to re-download your entire library.

+++ core/download_manager.lua
@@ -178,17 +178,19 @@ function DownloadManager.downloadDownloadList(browser)
 	end
 
 	local dl_count = #browser.downloads
-	for i = dl_count, 1, -1 do
-		local item = browser.downloads[i]
-		if downloaded and downloaded[item.file] then
-			table.remove(browser.downloads, i)
-		else -- if subprocess has been interrupted, check for the downloaded file
-			local attr = lfs.attributes(item.file)
-			if attr then
-				if attr.size > 0 then
-					table.remove(browser.downloads, i)
-				else -- incomplete download
-					os.remove(item.file)
+	if dl_count and dl_count > 0 then
+		for i = dl_count, 1, -1 do
+			local item = browser.downloads[i]
+			if downloaded and downloaded[item.file] then
+				table.remove(browser.downloads, i)
+			else -- if subprocess has been interrupted, check for the downloaded file
+				local attr = lfs.attributes(item.file)
+				if attr then
+					if attr.size > 0 then
+						table.remove(browser.downloads, i)
+					else -- incomplete download
+						os.remove(item.file)
+					end
 				end
 			end
 		end
@@ -234,6 +236,8 @@ function DownloadManager.downloadPendingSyncs(browser, dl_list)
 			return dl, dupe_list
 		end, info)
 
+		-- clear the dupe_list if this is a skip dups runs - we needed to weed them out, but we don't want to consider them
+		if browser.sync_skip then duplicate_list = {} end
 		if completed then
 			UIManager:close(info)
 		end
